Upddating exercise 6 that I forgot to save

+++ chapt3/exercise6.c
@@ -4,9 +4,9 @@ int main (void)
 {
    float x = 2.55, val;
 
-   float ex1, ex2;
-   ex1 =  x^3;
-   ex2 = x^2;
+   int ex1, ex2;
+   ex1 = (int) x^3;
+   ex2 = (int) x^2;
 
    val = 3 * ex1 - 5 * ex2 + 6;
